+++ Database/Migrations/2026_03_15_000002_create_scheduled_conversation_logs_table.php
@@ -12,7 +12,7 @@
  *
  * @package Modules\ScheduledConversations
  * @author  Raimundo Alba
- * @version 1.5.0
+ * @version 1.6.0
  */
 
 use Illuminate\Support\Facades\Schema;